Refactor navigation.
- Introduce SwitchNavigator for top level navigator.
- Manage loading routine in dedicated screen.
- Wrap all components in Redux.
- Delete useless Registry class.
