+++ src/navigation/account/AccountOrdersPage.js
@@ -8,36 +8,41 @@ import {
   Container,
   Header,
   Left, Right, Body,
-  Title, Content, Footer, Button, Icon, Text,
-  Label, Item, Input
+  Title, Content, Footer, Button, Icon, List, ListItem, Text
 } from 'native-base';
+import { connect } from 'react-redux'
 import { translate } from 'react-i18next'
+import { formatPrice } from '../../Cart'
 
-class AccountDetailsPage extends Component {
+class AccountOrdersPage extends Component {
   constructor(props) {
     super(props);
     this.state = {
       loading: true,
-      email: null,
-      username: null
+      orders: []
     };
   }
   componentDidMount() {
-    const { client } = this.props.navigation.state.params
-    client.get('/api/me')
-      .then(data => {
-        const { email, username } = data
+    this.props.httpClient.get('/api/me/orders')
+      .then((data) => {
         this.setState({
           loading: false,
-          email,
-          username,
+          orders: data['hydra:member']
         })
       });
   }
+  _renderRow(order) {
+
+    const { navigate } = this.props.navigation
+
+    return (
+      <ListItem onPress={() => navigate('OrderTracking', { order }) }>
+        <Body><Text>{ order.restaurant.name }</Text></Body>
+        <Right><Text>{ formatPrice(order.total) } €</Text></Right>
+      </ListItem>
+    );
+  }
   render() {
-
-    const { email, username } = this.state
-
     let loader = (
       <View />
     )
@@ -56,17 +61,8 @@ class AccountDetailsPage extends Component {
 
     return (
       <Container>
-        <Content style={{ paddingHorizontal: 10, paddingTop: 20 }}>
-          { username && (<Item stackedLabel disabled>
-            <Label>{this.props.t('USERNAME')}</Label>
-            <Input disabled placeholder={ username } />
-            <Icon name="information-circle" />
-          </Item> )}
-          { email && (<Item stackedLabel disabled>
-            <Label>{this.props.t('EMAIL')}</Label>
-            <Input disabled placeholder={ email } />
-            <Icon name="information-circle" />
-          </Item> )}
+        <Content>
+          <List dataArray={ this.state.orders } renderRow={ this._renderRow.bind(this) } />
         </Content>
         { loader }
       </Container>
@@ -83,4 +79,10 @@ const styles = StyleSheet.create({
   },
 });
 
-module.exports = translate()(AccountDetailsPage)
+function mapStateToProps(state) {
+  return {
+    httpClient: state.app.httpClient
+  }
+}
+
+module.exports = connect(mapStateToProps)(translate()(AccountOrdersPage))
