+++ src/navigation/navigators/index.js
@@ -1,0 +1,7 @@
+import HomeTab from './HomeTab'
+import AppStack from './AppStack'
+
+export default {
+  HomeTab,
+  AppStack
+}
